+++ tools/bamSitesToWig.py
@@ -3,7 +3,7 @@
 __author__ = ["Nathan C. Sheffield", "Jason Smith"]
 __credits__ = []
 __license__ = "BSD2"
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 __email__ = "nathan@code.databio.org"
 
 from argparse import ArgumentParser
@@ -92,48 +92,51 @@ class CutTracer(pararead.ParaReadProcessor):
         grab a subset of reads from the bamfile
         """
 
-        _LOGGER.debug(f"chrom: {chrom}")
         chrom_size = self.get_chrom_size(chrom)
 
         _LOGGER.debug("[Name: " + chrom + "; Size: " + str(chrom_size) + "]")
         reads = self.fetch_chunk(chrom)
 
-        chromOutFile = self._tempf(chrom.replace("|","_"))
-        chromOutFileWig = chromOutFile + ".wig"
+        chromOutFile = self._tempf(chrom)
         chromOutFileBw = chromOutFile + ".bw"
-        chromOutFileWigSm = chromOutFile + "_smooth.wig"
-        chromOutFileBwSm = chromOutFile + "_smooth.bw"
-        tmpFile = chromOutFile + "_cuts.txt"
 
         cutsToWig = os.path.join(os.path.dirname(__file__), "cutsToWig.pl")
 
         cmd1 = ("sort -n | perl " + cutsToWig + " " + str(chrom_size) +
-                " " + str(self.variable_step) + " " + str(self.scale) +
-                " > " + chromOutFileWig)
-        cmd2 = ("wigToBigWig -clip -fixedSummaries -keepAllChromosomes " +
-                chromOutFileWig + " " +
+                " " + str(self.variable_step) + " " + str(self.scale))
+        cmd2 = ("wigToBigWig -clip -fixedSummaries -keepAllChromosomes stdin " +
                 self.chrom_sizes_file + " " + chromOutFileBw)
         _LOGGER.debug("  cutsToWigProcess: " + cmd1)
         _LOGGER.debug("  wigToBigWigProcess: " + cmd2)
 
         if self.exactbw:
             cutsToWigProcess = subprocess.Popen(cmd1, shell=True,
-                stdin=subprocess.PIPE)
-            
+                stdin=subprocess.PIPE, stdout=subprocess.PIPE)
+            wigToBigWigProcess = subprocess.Popen(
+                ['wigToBigWig', '-clip', '-fixedSummaries',
+                 '-keepAllChromosomes', 'stdin',
+                 self.chrom_sizes_file, chromOutFileBw],
+                 stdin=cutsToWigProcess.stdout)
 
         if self.smoothbw:
             cutsToWigSm = os.path.join(os.path.dirname(__file__),
                                        "smoothWig.pl")
+            chromOutFileBwSm = chromOutFile + "_smooth.bw"
+            tmpFile = chromOutFile + "_cuts.txt"
             cmd1 = ("sort -n | tee " + tmpFile + " | perl " + cutsToWigSm +
                     " " + str(chrom_size) + " " +  str(self.smooth_length) +
                     " " + str(self.step_size) + " " + str(self.variable_step) +
-                    " " + str(self.scale) + " > " + chromOutFileWigSm)
+                    " " + str(self.scale))
             cmd2 = ("wigToBigWig -clip -fixedSummaries " +
-                    "-keepAllChromosomes " + chromOutFileWigSm + " " +
-                    self.chrom_sizes_file + " " + chromOutFileBwSm)
+                    "-keepAllChromosomes stdin " + self.chrom_sizes_file +
+                    " " + chromOutFileBwSm)
             cutsToWigProcessSm = subprocess.Popen(cmd1, shell=True,
-                stdin=subprocess.PIPE)
-            
+                stdin=subprocess.PIPE, stdout=subprocess.PIPE)
+            wigToBigWigProcessSm = subprocess.Popen(
+                ['wigToBigWig', '-clip', '-fixedSummaries',
+                 '-keepAllChromosomes', 'stdin',
+                 self.chrom_sizes_file, chromOutFileBwSm],
+                 stdin=cutsToWigProcessSm.stdout)
 
         if self.bedout:
             chromOutFileBed = chromOutFile + ".bed"
@@ -215,10 +218,10 @@ class CutTracer(pararead.ParaReadProcessor):
                 shifted_pos = get_shifted_pos(read, shift_factor)
 
                 if self.exactbw and shifted_pos:
-                    cutsToWigProcess.stdin.write((str(shifted_pos) + "\n").encode('utf-8'))
+                    cutsToWigProcess.stdin.write((str(shifted_pos + 1) + "\n").encode('utf-8'))
 
                 if self.smoothbw and shifted_pos:
-                    cutsToWigProcessSm.stdin.write((str(shifted_pos) + "\n").encode('utf-8'))
+                    cutsToWigProcessSm.stdin.write((str(shifted_pos + 1) + "\n").encode('utf-8'))
 
                 if self.bedout and shifted_pos:
                     strand = "-" if read.is_reverse else "+"
@@ -237,13 +240,8 @@ class CutTracer(pararead.ParaReadProcessor):
             # Clean up processes
             if self.exactbw:
                 cutsToWigProcess.stdin.close()
-                cutsToWigProcess.communicate()
                 _LOGGER.debug("Encoding exact bigwig for " + chrom + 
                               " (last read position:" + str(read.pos) + ")...")
-                wigToBigWigProcess = subprocess.Popen(
-                ['wigToBigWig', '-clip', '-fixedSummaries',
-                 '-keepAllChromosomes', chromOutFileWig,
-                 self.chrom_sizes_file, chromOutFileBw])
                 wigToBigWigProcess.communicate()
 
             if self.bedout:
@@ -251,13 +249,8 @@ class CutTracer(pararead.ParaReadProcessor):
 
             if self.smoothbw:
                 cutsToWigProcessSm.stdin.close()
-                cutsToWigProcessSm.communicate()
                 _LOGGER.debug("Encoding smooth bigwig for " + chrom +
                               " (last read position:" + str(read.pos) + ")...")
-                wigToBigWigProcessSm = subprocess.Popen(
-                ['wigToBigWig', '-clip', '-fixedSummaries',
-                 '-keepAllChromosomes', chromOutFileWigSm,
-                 self.chrom_sizes_file, chromOutFileBwSm])
                 wigToBigWigProcessSm.communicate()
 
         except StopIteration as e:
@@ -277,12 +270,11 @@ class CutTracer(pararead.ParaReadProcessor):
             _LOGGER.info("No successful chromosomes, so no combining.")
             return
         elif len(good_chromosomes) == 1:
-            subprocess.call(["mv",
-                            self._tempf(good_chromosomes[0].replace("|","_")) +
+            subprocess.call(["mv", self._tempf(good_chromosomes[0]) +
                              ".bw", self.exactbw])
-            subprocess.call(["mv",
-                            self._tempf(good_chromosomes[0].replace("|","_")) +
+            subprocess.call(["mv", self._tempf(good_chromosomes[0]) +
                              "_smooth.bw", self.smoothbw])
+
         else:
             if self.exactbw:
                 _LOGGER.info("Merging {} files into output file: '{}'".